+++ backend/src/main/java/ai/verta/modeldb/experimentRun/FutureExperimentRunServiceImpl.java
@@ -16,7 +16,7 @@ import ai.verta.modeldb.common.exceptions.ModelDBException;
 import ai.verta.modeldb.common.exceptions.NotFoundException;
 import ai.verta.modeldb.common.futures.FutureGrpc;
 import ai.verta.modeldb.common.futures.InternalFuture;
-import ai.verta.modeldb.exceptions.InvalidArgumentException;
+import ai.verta.modeldb.common.exceptions.InvalidArgumentException;
 import ai.verta.uac.GetResourcesResponseItem;
 import com.google.gson.Gson;
 import com.google.gson.JsonElement;
