+++ backend/src/main/java/ai/verta/modeldb/experimentRun/subtypes/SortingHandler.java
@@ -5,7 +5,7 @@ import ai.verta.modeldb.common.futures.InternalFuture;
 import ai.verta.modeldb.common.query.OrderColumn;
 import ai.verta.modeldb.common.query.OrderTable;
 import ai.verta.modeldb.common.query.QueryFilterContext;
-import ai.verta.modeldb.exceptions.InvalidArgumentException;
+import ai.verta.modeldb.common.exceptions.InvalidArgumentException;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
